+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Receiving app closes after closing Routing list page. Refs UIREC-380.
 * Improve item record behavior for items associated with receiving pieces after binding. Refs UIREC-371.
 * Move the piece form out of the modal to a separate route with a full view. Refs UIREC-368.
+* Clear Item details when "Save and create another" option for adding electronic piece. Refs UIREC-377.
 
 ## [5.0.4](https://github.com/folio-org/ui-receiving/tree/v5.0.4) (2024-04-22)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v5.0.3...v5.0.4)